pyproject.toml
[tool.poetry]
name = "web-gears"
version = "0.1.0"
description = ""
authors = ["vchernavskiy <vladimir>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-jwt = "^1.11.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest-django = "^4.9.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
